agregue skills

+++ src/main/java/com/portfolio/rbl/Controller/PersonaController.java
@@ -34,14 +34,14 @@ public class PersonaController {
     }
        @PreAuthorize ("hasRole('ADMIN')")
     @DeleteMapping ("/personas/borrar/{id}")
-    public String deletePersona(@PathVariable Long id){
+    public String deletePersona(@PathVariable int id){
     ipersonaService.deletePersona(id);
     return "La persona fue eliminada correctamente";
     }
     
        @PreAuthorize ("hasRole('ADMIN')")
     @PutMapping("/personas/editar/{id}")
-    public Persona editPersona(@PathVariable Long id,
+    public Persona editPersona(@PathVariable int id,
                                 @RequestParam ("nombre") String nuevoNombre,
                                 @RequestParam ("apellido") String nuevoApellido,
                                 @RequestParam ("img") String nuevoImg){
@@ -57,6 +57,6 @@ public class PersonaController {
     
     @GetMapping ("/personas/traer/perfil")
     public Persona findPersona() {
-    return ipersonaService.findPersona((long)1);
+    return ipersonaService.findPersona((int)1);
             }
 }
