+++ src/main/java/com/portfolio/rbl/Interface/IPersonaService.java
@@ -1,3 +1,4 @@
+
 package com.portfolio.rbl.Interface;
 
 import com.portfolio.rbl.Entity.Persona;
@@ -5,15 +6,11 @@ import java.util.List;
 
 
 public interface IPersonaService {
-    //Traer una persona
     public List<Persona> getPersona();
-   
-    //guardar un objeto de tipo persona
+    
     public void savePersona(Persona persona);
     
-    //eliminar un usuario buscar por id
-     public void deletePersona(Long id);
-     
-     //buscar persona por id
-     public Persona findPersona(Long id);
+    public void deletePersona(int id);
+    
+    public Persona findPersona(int id);
 }
