+++ src/main/java/com/portfolio/rbl/Service/ImpPersonaService.java
@@ -8,14 +8,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
+
 public class ImpPersonaService implements IPersonaService{
-    @Autowired IPersonaRepository ipersonaRepository;
-    
-    
+@Autowired IPersonaRepository ipersonaRepository;
+
     @Override
     public List<Persona> getPersona() {
-        List<Persona> persona = ipersonaRepository.findAll();
-                return persona;
+    List<Persona> persona = ipersonaRepository.findAll();
+    return persona;
     }
 
     @Override
@@ -24,14 +24,15 @@ public class ImpPersonaService implements IPersonaService{
     }
 
     @Override
-    public void deletePersona(Long id) {
-        ipersonaRepository.deleteById(id);
+    public void deletePersona(int id) {
+       ipersonaRepository.deleteById(id);
     }
 
     @Override
-    public Persona findPersona(Long id) {
+    public Persona findPersona(int id) {
         Persona persona = ipersonaRepository.findById(id).orElse(null);
         return persona;
     }
+
     
 }
